Refactor to allow API usage in DetailView
